stop loss.

+++ preframr/regdataset.py
@@ -65,7 +65,9 @@ class RegDataset(torch.utils.data.Dataset):
                     seq = None
                     if self.tokenizer.tokens is not None:
                         df = self.tokenizer.merge_token_df(self.tokenizer.tokens, df)
-                        seq = self.tokenizer.encode(df["n"].astype(np.int64).to_numpy())
+                        seq = self.tokenizer.encode(
+                            df["n"].astype(np.int64).to_numpy()
+                        ).astype(np.int64)
                         if len(seq) < self.args.seq_len:
                             self.logger.info(
                                 "rejecting sequence from %s too short %u",
